--- module_4/common.php
@@ -1,6 +1,0 @@
-<?php
-/**
- * Escapes HTML for output
- *
- */
-
